removes make_* from test code.
This code I believe was not operating correctly and potentially
accessing memory in an illegal way. Specifically, it allocated
a zero byte struct and returned a pointer to said struct.
These zero byte structs were forward declarations, so the correct
way to use them are as direct pointers that get "filled" in by
IonC.

I updated the tests to just initialize them as null pointers.