Initial pure-Rust impls for Decimal and Timestamp
When possible, Decimal uses a u64 to represent its coefficient. If
necessary, it can fall back to a BigUint.

When possible, Timestamp uses a u32 field to store a number of
nanoseconds for its fractional precision. If necessary, it can
fall back to a Decimal value.

The Timestamp type includes a first pass at a builder API that
takes care to prevent the user from accidentally constructing
an invalid Timestamp. Fields must be specified in increasing
order of precision. Hour and minute must always be specified
together. Offsets may optionally be specified for Timestamp
precisions that include a time.