Adds a preliminary `Element` DOM-like API.
Provides an _borrowed_/_owned_ implementations of the traits in the
`value` module.  The borrowed implementation is tied to a lifetime for
string/binary data (but potentially more in the future).  The owned
implementation is completely self contained.

* Defines `SymbolToken` and `ImportSource` traits that is the basis for
  full symbol/field name/annotation support in `ion-rs`, but also is
  required for the `Element` trait.
* APIs that return `Iterator` types currently use a `Box` to avoid the
  need for generic associated types (GAT).
* Captures the variance of integers in Ion with the `IntAccess` trait
  and the `AnyInt` enum.
* Currently implements the `Struct` trait via `Vec` of tuples, which
  will have to be refactored for O(N) access but is suitable for the
  iterator APIS.

Starts to address #114, but is not nearly complete.  This commit is to
provide the initial structure of this API for collaborative development,
and to get early review.