Refactors `value::owned` unit tests to be more parameterized.
Increases coverage for that module from 54% to 72%.

Creates a dispatch table style approach where all the `asXXX` methods
are negatively asserted and each test case "overrides" the negative
assertion cases with its own.

Also adds `PartialEq`/`Eq` for `AnyInt` to facilitate testing.

Addresses some of #146 and is based on work that @desaikd started with #147.