Implement Ion Hash for Strings with sha256
This commit adds an `IonHasher<D: Digest>` struct that can produce a
digest for a given `OwnedElement`.

The `Digest` trait comes from the `sha2` crate, which is absolutely not
what we want! However, I figured this was the shortest path to getting
something working.

There is one test, written by hand. We'll integrate with the ion-hash
test suite shortly.

Future work will:

- Flesh out the implementation for other element types
- Abstract over the hasher
- Provider Reader/Writer types that hash-as-you-go
- Abstract over Element