Start to wire up ion-hash-tests
This commit starts to implement a checker for the cases described in
ion_hash_tests.ion. For now, we support only:

- Ion text
- identity with update, update, ... digest

The actual file isn't being loaded yet, because it fails to parse. So, I
just copied in the first test case by hand. This let me start to learn
the case structure without getting bogged down by the parse failure.

The first test was against nulls, so I had to implement that in the
hasher! It also made me make the `new` and `hash_element` fns public..
doh!

To facilitate testing, we now decouple from the `sha2::Digest` trait and
have our own (worse..) trait. We'll make it better over time; for now I
just got something in place.

The actual test has a simple hand-written stub. It uses interior
mutability so we can easily spy on what method calls it received.