Support hashing bools, inline a subset of the suite
This commit does 3 things:

1. The internal hash_TYPE methods now call the hasher directly. This
cuts out a bunch of allocations.

2. hash_bool is now implemented.

3. We start to use the test suite directly..

.. by copying `ion_hash_tests.ion` over from the gitsubmodule and
pruning it by hand. The pruning achieves two things:

1. It only includes tests that pass
2. The entire file didn't _parse_ (see #219)

Ideally, we can work around point 1 with a skip list, however.. not all
the tests are named yet. There is implicit naming by
Ion-text-representation, but we don't yet have a dumper!