Initial Ion Hash Testing (#227)
This commit starts to implement a checker for the cases described in
ion_hash_tests.ion. For now, we support only:

- Ion text
- identity with update, update, ... digest

The actual file isn't being loaded yet, because it fails to parse. So, I
just copied in the first test case by hand. This let me start to learn
the case structure without getting bogged down by the parse failure.

The first test was against nulls, so I had to implement that in the
hasher! It also made me make the `new` and `hash_element` fns public..
doh!

To facilitate testing, we now decouple from the `sha2::Digest` trait and
have our own (worse..) trait. We'll make it better over time; for now I
just got something in place.

The actual test has a simple hand-written stub. It uses interior
mutability so we can easily spy on what method calls it received.

After chatting with @almann, we agreed to skip the intermediate
assertions (calls to `update`) and focus on the final digest values. We
currently believe that the intermediate checks are only useful for
debugging and not needed for correctness. Not having these checks makes
the tests simpler.

Additional notes:
* Adds ion-hash-test as a submodule.
* Remove hex-literal dependency.
* Use digest crate for the Digest trait.
* Support hashing `bool`.
* The internal hash_TYPE methods now call the hasher directly. This
  cuts out a bunch of allocations.

Co-authored-by: Marc Bowes <bowes@amazon.com>