Adds the `Dumper` API.
Provides a fixed buffer output implementation using `ion-c-sys`.
This implementation uses very specific `unsafe` code to return
the written data as an output slice.  It has to do this because
`IonCWriterHandle` borrows the input slice to write *implicitly*,
though the lifetime is tracked through the wrapper.

* Adds basic configuration enums.
* Adds constraint for `Builder<SymbolToken = Self::SymbolToken>`
  for `Element`.
* Adds `Into<ION_TYPE>` for `IonType`.
* Adds `try_to!` macro to integrate `Option` with `IonResult`
  more easily (h/t to @marcbowes in #215)
* Adds an doc test to demonstrate usage in the main module.
* Adds basic unit tests that have expectation on encoded bytes.
* Adds round trip testing for `ion-tests` and `good` tests for various
  permutations of loading/dumping values.  More sophisticated
  testing with `equivs` and `non-equivs` can be done, but that is
  left as a TODO (#236)
  - Adds skip list for issues with the round-tripping (#235).

Note that `try_dumper_for_slice` returns a concrete type instead of an
opaque `impl Dumper<Output = &[u8]>` because writing higher-order
functions that construct the above type appears to require GAT (HRTB
cannot be applied to the associated type to avoid this).  As a result,
the `IonCWriterFixedDumper` type and an alias of `FixedDumper` is public
(unlike the corresponding types in `Dumper`).

Resolves #197.