Version bump for all crates to publish. (#255)
This gets `ion-rs` ready to publish for new APIs.

Patch level version bump for `ion-hash` and `ion-c-sys`.

Made the dependencies for all crates consistent.  Notably changed the
dependencies to `x.y` format over `x.y.z`.  The rationale for this
change is that we've been a bit rigid on patch versions and that has
caused some downstream compatibility woes.  If any particular crates we
depend on break semver, we can always add the patch version in for those
crates (and/or use tilde versions).

Another change is switching from `^x.y` to just `x.y` as these syntaxes
are the same as per the Cargo book and we had a mixture of both, so I
went with preferring the one that crates.io naturally suggests.

https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html

Also updates code coverage CI to `nightly-2021-05-07`.