Refactor ion-hash implementation (#264)
Squash of 3 changes:

* Refactor IonTypeCode

Use TryFrom for u8 -> IonTypeCode and IonTypeCode -> IonType
conversions. This is idiomatic in Rust.

IonTypeCode is now pub (was: `pub(crate)`). The thinking is that I'd
like to use these typecodes from ion-hash (which is a different crate).

* Refactor ion-hash to use IonTypeCode

This commit exposes IonTypeCode from ion-rust so that ion-hash can use
it to compute TypeQualifiers.

TypeQualifiers are now computed using `const fn`, hopefully in a way
that is much easier to understand than the spec or other
implementations!

* Add int support

The final int test (that tests BigInt) is elided because the digest is
too long. Next commit I'll move back to a cloned test digester.