TestDigest now uses interior mutability
The fixed-length output requirements of the Digest trait make
integrating with the ion-hash-test suite a little annoying! Nothing
wrong with the trait; rather it's a consequence of the design of the
test suite..

The tests define correct as "the underlying digest was fed a series of
updates that look like this" - kind of like a mock/expect setup. Because
different test cases have different sequences, this design clashes with
the fixed length expectations of the Digest trait.

To work around this, we completely ignore the fixed-length result in our
tests and return assert on a result vec that lives outside of the Digest
trait implementation.

One wrinkle is that IonHasher wants to own the Digest so that it can be
sure nobody else is writing to it. To work around that, we use interior
mutability to allow our mock to be cloned and our tests keep a concrete
reference to it.

As a direct consequence of this change, I can enable the intLength512
test which was longer that our previous 256 bit length limit. Also, the
nonsense around trailing zeros is gone!