Refactor out `raw::i64_magnitude` (#266)
Refactor writing i64 magnitude subfield

This commit refactors the code that produces the "magnitude" subfield
out of `BinarySystemWriter`, allowing the code to be re-used in
ion-hash.

The code is moved into `uint.rs` and a duplicate copy is removed as a
result. 

ion-hash now avoids i64 -> BigInt as a result of this change.