Rework ion-hash representations (#268)
Before this commit, every `Element` visited during ion hash computation
would need to allocate a vector: because each element type can be
represented by a variable number of bytes! For example, a boolean has no
representation while a string is representated by its UTF8 bytes.

This commit fixes that by having the representation implementation write
to the digester directly. For some types (like a BigInt) we will still
need a variable length vector. But in other cases, we now no longer need
to allocate.

While implementing this, I wanted to make the escape method also not
require allocation. This is a simple case of wrapping the `update`
method to maybe write two bytes per byte written. However, because we
were dependent on the `Digest` trait, writing the wrapper required
wrapping all of those methods. In this commit, we move away from the
`Digest` trait and rather use the traits that *make up* that trait. This
means the escaping wrapper just needs to worry about `update`.

Co-authored-by: Almann Goo <almann.goo@gmail.com>