Binary timestamp is now precision aware
This commit moves binary encoding off the Chrono type and onto the
in-crate Timestamp type. Importantly, this is precision-aware and we now
elide subfields below the intended precision. For example:

```
`2021-05T`
```

.. used to be encoded all the way down to nanosecond precision, but now
will stop after encoding offset, year and month (as defined by the
spec).

Unfortunately the Timestamp type doesn't allow for borrowed Chrono
types, so we added a clone in the binary writer. We'll have to fix that
up at some point, but it's not clear to me right now that the API on
writer (`write_datetime`) is even the one we'll end up with!