Implement IonHash support for structs (#270)
Quite a lot going on in this commit! Structs are different from the
scalars we've been hashing so far in that they create **new hashers**:

> escape(concat(sort(H(field1), H(field2), ..., H(fieldn))))

The big `H` there is an actual hash computation.

This immediately means that `IonHasher` not be a useful abstraction
since the requirement to be able to make new hashers means that asking
for the initial hasher is a little silly. So, in this commit I stripped
that struct away. This creates a bit of weirdness in lib.rs where there
are a bunch of functions that take a mutable hasher as their second
argument. My intention is to add some other abstraction (maybe a trait)
in a future commit. [fn:1]

The next bump in the road was that all our abstractions coupled
"dispatch" and "implementation", where the former means "look at the Ion
Type then do something specific for that type". This posed a problem
when computing the field hash: the key (of a field) is a Symbol and a
Symbol is *not an Element*. To fix this, I decoupled dispatch and
implementation, allowing the field hash computation to go directly to
the TQ/repr infrastructure for symbols.

The biggest bump by *far* was yet more trouble with the ion-hash-test
suite. Using a real hasher (sha256) became extremely tricky with the
field hashers because the expectations don't show that. For example:

```ion
{
  ion:{'':1},              // empty symbol
    // ^^^^ NOTE: this field requires a *new hasher*
  expect:{ identity:(
    update::(0x0b) update::(0xd0)
    // --------LOOK DOWN! -------------------
        update::(0x0b) update::(0x70) update::(0x0e)
        update::(0x0b) update::(0x20) update::(0x01) update::(0x0e)
        digest::(0x0b 0x70 0x0e 0x0b 0x20 0x01 0x0e)
    // ^ this idented sequence is on a new instance of the hasher
    // --------LOOK UP! ---------------------
      update::(0x0c 0x0b 0x70 0x0c 0x0e 0x0c 0x0b 0x20 0x01 0x0c 0x0e)
    update::(0x0e)
    digest::(0x0b 0xd0 0x0c 0x0b 0x70 0x0c 0x0e 0x0c 0x0b 0x20 0x01 0x0c 0x0e 0x0e)
  ) }
}
```

So, I've gone back to an `IdentityHasher` with fixed size output and
worked around the trailing zeros.

A few (3) of the struct tests fail because structs also test other types
that we haven't implemented. I added an ignore list so that we can have
all the struct tests in this file and clearly show which aren't yet
expected to pass.

[fn:1]: The struct representation has some duplicate sequence code that
I plan to fix at the same time.

Co-authored-by: Zack Slayton <zack.slayton@gmail.com>