Wire up ion-hash for decimals (#285)
This commit uses the newly extracted binary decimal encoder to support
ion-hash for timestamps. We fix two bugs related to
representation (positive zero & eliding +0 coefficients).

Unfortunately, negative zero coefficients had to be commented out. The
issue is described in #83. The bug is such that I couldn't even skip it
since the loader mangles the "name".

Co-authored-by: Almann Goo <almann.goo@gmail.com>