Move ion-hash implementation behind traits
Before this commit, we were passing around a sink (that implements
digest or write) as a `&mut` "out" parameter. Ugly in general, but even
more so given we had to repeat the `Digest` subtraits over and over.

Now, we have two traits (split simply for code organization). One trait
handles the overall hashing orchestration, while another handles the
representation. Other than some funny business with calling syntax, this
worked out pretty nicely!

Note that I absolutely have _not_ thought through the traits
nicely (i.e. "are we happy with the public/private split?" was not a
consideration). Rather, I focused on the mechanical refactoring to see
if this was worth it. There are loads of doc/comments that don't make
sense anymore that I'll fix up shortly!