Return bytes written for timestamp/decimal encoding (#289)
This commit update `VarInt::write_i64` to return number of bytes
written, and then weaves that information through the timestamp
and decimal encoders. Some tests were added.

Some `#[rustfmt::skip]` annotations are moved from methods to
variables. The skip is useful because the "art" of the declaration has
value, however, applying the skip to the entire method was overkill.