Makes IonDataSource report EOF reliably
Prior to this commit, multiple methods in the IonDataSource trait would
either loop infinitely or return an `IoError` in the event of an
unexpected EOF.

This commit causes all of the methods in question to detect EOFs and
report them as a decoding error. Work needed to add a proper
IonError::IncompleteData enum variant is being tracked in #299.

Fixes #300.