Fixes panic in read_datetime (#310)
The binary cursor's `read_datetime` method incorrectly 0-indexes
its default month and day values. If a timestamp is encountered
that only specifies a year or year+month, the default value of
zero causes the construction of a DateTime to panic.

This patch sets the default month and day values to 1.

Fixes #306.