Retains fractional secs when converting to DateTime
Prior to this patch, Timestamps that represented their fractional
seconds as a Decimal would lose their subsecond precision when
converting to a NaiveDateTime or DateTime<FixedOffset> via
`try_into()`.

This patch causes the decimal fractional seconds to be converted
to a number of nanoseconds. This version only converts decimals
that use a u64 magnitude, which is the most common case. We can
and should implement this for BigUInt magnitudes in a future change.