Adds `read_timestamp` method to the binary reader (#312)
This PR adds a `read_timestamp` method alongside the existing
`read_datetime` method, which is now marked deprecated. This
allows users to read an Ion timestamp with full fidelity.

It also adds implementations of `TryInto` to allow users to
convert a Timestamp into the corresponding chrono type:
* Timestamps with unknown offsets can be turned into
  `NaiveDateTime`s.
* Timestamps with known offsets can be turned into
  `DateTime<FixedOffset>`s.

In order to report whether a given Timestamp has a known or
unknown offset, the `VarInt` type now stores an extra bool
field to track whether it is actually negative zero. Prior
to this commit, the value of the `VarInt` was stored as an
i64, which cannot represent the -0 case on its own.

Finally, this PR fixes a bug in the way that Timestamps with
different Mantissa representations (Decimal vs number-of-digits)
were being tested for equality.