Adds streaming support to the text reader
This PR allows a `TextReader` to be created from a variety of text
data sources, including `String`, `&str`, `&[u8]`, and `BufReader`.
Support for other types can be addede by implementing the
`TextIonDataSource` trait.

It also introduces the `TextBuffer`, a type which can pull text
into a buffer `n` lines at a time and discard bytes as they're
processed. This enables the reader to process its input in a
streaming fashion, holding a modest amount of data in memory at
a time.

THe patch adds parsing logic to detect the beginnings and endings of
the various container types. However, it does not add logic for
stepping into/out of those containers.

Finally, by leveraging the above features, the existing
proto-TextReader can now iterate over a text Ion stream containing
multiple top-level values. If it encounters an error, it can report
the line of input that caused it.