Resolves ambiguous text parsing at EOF (#320)
This PR modifies the text reader to be able to unambiguously
parse data found at the end of the stream. Because the
parser always operates on a fixed buffer, it could not easily
distinguish between the end of the buffer (more data to come)
and the end of the stream being loaded into the buffer (EOF).

When EOF is detected, the reader will now append a sentinel
value to the end of the stream and re-attempt parsing. If the
sentinel value is found, `EndOfStream` is reported. If a different
value is found, that value is returned instead and the sentinel
is discarded.

Fixes #318. See that issue for more details.