Implements the `RawReader` trait for the `RawTextReader` (#336)
This PR renames a number of constructs:

* The `Cursor` trait is now the `RawReader` trait to better align with
  terminology used in other libraries.
* The `TextReader` is now the `RawTextReader` as it does not provide
  any symbol management.
* The `BinaryIonCursor` (which implements `RawReader`) is now the
  `RawBinaryReader`.

Prior to this patch, the `RawReader` trait used `SymbolId` as the
return type for methods that read annotations, struct field names,
and symbols because that was/is the only kind of symbol token
that appears in binary. Now that we have a `RawTextReader`, we
also need to be able to read symbol tokens that define their text
inline. This change set introduces a `RawSymbolToken` enum that
can store either a `SymbolId` or a `String`. The `RawReader` trait
has been updated to use `RawSymbolToken` throughout.

Finally, this PR implements `RawReader` for the `RawTextReader`,
making it possible to navigate an Ion text stream using the expected
`next()`, `step_in()`, `step_out()`, and `read_TYPE()` method calls.