reexport bigdecimal
This commit pub-uses bigdecimal so that consumers can use the bigdecimal
type without have explicit exact-version dependency.

bigdecimal is a <1.0 library, and so cargo will pull in additional
copies of it by default, as semver says each release of a <1.0 crate is
backwards incompatible. So, if a consumer of ion-rust didn't have the
exact right version, then the `TryFrom<Decimal> for BigDecimal`
implementation wouldn't be accessible.

Fixes #302