System reader (#344)
This PR defines a new type of streaming reader called a `SystemReader`.
The `SystemReader` processes 'system' elements that it encounters in the
Ion stream, but it does not filter them out. This allows applications to
visit Ion values used to encode local symbol tables without also being
responsible for interpreting them.

The `SystemReader` tracks its position within the current LST (if any)
each time that it advances with `next()`, `step_in()`, and `step_out()`.
If the application skips over some or all of a local symbol table, the
`SystemReader` will automatically process the rest of the LST before
advancing.

The `SystemReader` wraps a provided implementation
of the `RawReader` trait, allowing it to be format-agnostic.

This PR also removes the `SystemEventHandler` API from the user-level
`Reader`. This was previously used to allow low-level tooling a way to
react to system values in the stream after they were processed, but the
information it exposed was limited. Low-level tools should use a
`SystemReader` instead, allowing them to process system-level elements
the same way they would process user-level elements.

Other changes:
* Renames the `StreamItem` enum to `RawStreamItem`
* Adds a new `SystemStreamItem` enum
* `SymbolTable` is now capable of assigning symbol IDs unknown text