Adds user-level `TextWriter` and `BinaryWriter`
This PR introduces user-level wrappers for the `RawBinaryWriter`
and the `RawTextWriter`. These higher-level types manage a
symbol table and so can convert between text and symbol IDs,
adding new symbols when appropriate.

This PR also renames the `RawWriter` trait to `Writer`, as the
methods in its interface are applicable to both the raw and user
views of data.

While this PR adds basic unit test for the new wrapper types, it
stops short of providing exhaustive correctness checking. One or
more follow-on PRs will add full integration with ion-tests,
demonstrating spec conformance.