Adds `Symbol` type, `StreamReader` trait (#366)
* Adds `Symbol` type, `StreamReader` trait

This PR makes the following high-level changes:

1. The `SymbolTable` has been modified to store `Rc<str>` text
values instead of `String`s, allowing the text they represent
to be shared cheaply.
2. It introduces a new `Symbol` type that represents the resolved
text of a field name, annotation, or symbol value.
3. It adds a unifying `StreamReader` trait implemented by all
reader types.
4. It removes methods that worked with the external `DateTime` and
`BigDecimal` types. These methods were deprecated in v0.6.1.

**Changes to the `SymbolTable`**

The `SymbolTable` has been modified to store `Rc<str>` entities
that can be cheaply cloned and passed around rather than `String`s
that require more heap allocations to copy/share.

**The `Symbol` data type**

A `Symbol` represents the text of a fully resolved Ion symbol, field
name, or annotation. It is distinct from the various 'token' types in
that it does not expose information about how the symbol was originally
encoded in the stream. It aims to provide an ergonomic,
application-level text type with a static lifetime.

If the reader encounters a text symbol token in the stream, a `Symbol`
can be constructed with that token's text stored as a `String`.

If the reader encounters a symbol ID token in the stream, a `Symbol`
can be constructed with an `Rc<str>` pointing to that ID's corresponding
text in the `SymbolTable`.

**The `StreamReader` trait**

`StreamReader` is a trait that captures the format-agnostic
functionality needed to navigate through an Ion stream at any level of
abstraction.

`StreamReader` has two associated types:
1. `Item`: a type that represents the next entity found in the
   stream being read. This varies according to the implementor's
   level of abstraction.
2. `Symbol`: a type that represents a field name, annotation, or
   Ion symbol value at this level of abstraction. A `RawReader`
   would return a `RawSymbolToken` (a `SymbolId` OR `String`),
   allowing for the possibility that the symbol's text still needs
   to be resolved. In contrast, a `Reader` would return a fully
   resolved `Symbol` (a `String` or `Rc<str>` from the symbol table).

`StreamReader` is implemented by all of the following readers:
   * `RawBinaryReader`
   * `RawTextReader`
   * `SystemReader`
   * `Reader`

What was previously the `RawReader` trait is now an alias for
`StreamReader<Item=RawStreamItem, Symbol=RawSymbolToken>`.

The `StreamReader` changes the return type of most methods in the old
reader implementations from `IonResult<Option<T>>` to simply
`IonResult<T>`. This simplifies usage in the 'happy path' substantially.

We previously used `Option::None` to represent no-more-values or to
indicate that the current value could not be read with a given `read_*`
method. (e.g. An `int` could not be read with `read_string`.)
No-more-values is now represented by a new enum variant called `Nothing`
in the `StreamItem`, `SystemStreamItem`, `RawStreamItem` enums. Calling
a `read_*` method with an expected type that does not match the current
value's type will result in an `IonError::IllegalOperation`.

All `StreamReader` methods that require symbol resolution (`field_name()`,
`annotations()`) now return an `IonResult<Self::Symbol>` to allow for
the possibility of an `IonError::DecodingError` caused by undefined
text. Previously, such an error would cause a `panic!`.