Adds custom `IonParseResult` type (#379)
The `nom` crate's out-of-the-box default error type is intentionally
minimalist. Constructing and processing a `nom::error::Error` is
trivial; unfortunately, it also has limited in what it information it
can communicate to the user.

With this in mind, all of `nom`'s parsers are generic over not only
their input and output types, but also their error type. This allows
users to replace `nom::error::Error` with the type of their choice.

This PR introduces a new type called `IonParseError` that can store
additional information about problems that were encountered during
parsing. In the typical use case (unmatched text), it is as cheap to
construct and process as `nom::error::Error`. However, when a fatal
error is encountered (for example: an I/O error, overflow, etc), it can
store additional, heap-allocated context to report to the end user.

To simplify type signatures in the `text::parsers` module, a new
type alias for `IResult<&str, T, IonParseError>` has also been
added: `IonParseResult<T>`. This eliminates substantial visual
noise from the parsers.

Finally, two extension traits have been added to the library that
make it easier to mix `nom` parsers that work with `IResult` with
`ion-rs` parsers that work with `IonParseResult`.

Now that our parsers return `IonParseResult`, it is possible for us to
surface fatal errors to the application with valuable context. Many
situations that would previously lead to a panic now result in a
user-manageable `IonParseResult::Err`.

See the documentation for each type and trait in
`src/text/parsers/parse_result.rs` for more information.