Adds IonEq trait, fixes float equivalence (#382)
This PR adds the `IonEq` trait, which types can implement if their
implementation of `PartialEq` tests for a different definition of
equivalence than that of the Ion data model. Types that require
this include `float`, `timestamp`, and `decimal`.

The patch also modifies the `ion-tests` integration to use `ion_eq` (the
sole method of the `IonEq` trait) instead of `==` when doing roundtrip
testing or comparing equivs/non-equivs.