Adds `NativeElementWriter`, string/clob validation, big `DecodedInt`s (#384)
This PR makes the following changes:

* A native Rust `ElementWriter` implementation (`NativeElementWriter`)
has been added with basic testing; it is not yet being used in
`ion-tests`.
* ion-tests integration has been refactored to facilitate running
roundtrip tests with either the `IonCSliceElementWriter` or the
`NativeElementWriter` as desired.
* Fixed a bug in the ion-tests integration that caused the ion-c reader
to be used in some tests even when the native reader had been requested.
This change surfaced some additional failing ion-tests that have been
addressed in the changes below.
* The 64-bit size limit on the binary encoding primitive `Int` has been
lifted. This allows Timestamps and Decimals of arbitrary precision to be
read; write support for big `Int`s has not been added yet.
* Clobs now disallow unescaped non-ASCII characters and unescaped ASCII
control characters.
* Strings now disallow unescaped ASCII control characters.
* Strings and clobs now correctly decode utf-16 surrogate pairs.
* Long-form clobs and strings now normalize unescaped newlines.
* Fixed a bug in the `TextBuffer` that could cause a panic if restacking
the buffer's contents caused `truncate` to evaluate garbage data beyond
the truncation point.
* Fixed a bug in the way that Timestamp fractional seconds were being
converted to a Decimal.