Implements IonEq for Timestamp, Decimal (#387)
Timestamp equality (using `PartialEq` or `==`) now checks to see if the
two `Timestamp`s represent the same instant in time after applying their
respective offsets. Precision is ignored. The original field-by-field
comparison that considers precision has been moved to an implementation
of the `IonEq` trait.

Decimal equality (using `PartialEq` or `==`) now checks to see if the
two `Decimal`s represent the same mathematical value, ignoring
precision. In particular, positive and negative zeros of any exponent
are considered equali. The original comparison that requires the sign,
magnitude, and exponent to all match exactly has been moved to an
implementation of the `IonEq` trait.`