Reader/Writer builders, pretty text writer (#388)
This PR introduces (barebones) builder APIs for readers and
writers of different abstraction levels. This will allow us to add
new configuration elements over time in a backwards-compatible way.

It also adds a way to construct a "pretty" `TextWriter` that adds
generous, human-friendly whitespace to its output. This new version of
the `TextWriter` has been added to the `ion-tests` alongside the
standard one.

Finally, the `Reader` type has been renamed to `UserReader` to better
align with the names for the `Raw*Reader` and `SystemReader` types. The
name `Reader` is now an alias for a `UserReader<Box<dyn RawReader>>`,
which is the type of `Reader` we expect users to create the most often.
A `ReaderBuilder` will detect whether the provided data is text or
binary and return a `Reader` that abstracts over the format
differences via `Box<dyn RawReader>`.