Adds a non-blocking raw binary reader
The existing `RawBinaryReader` wraps an `io::BufRead` implementation
and pulls additional bytes from it as needed. This allows it to read
large files in a streaming fashion; however, it also means that the
file must be complete. If the data source is still growing (for
example: a file that is being appended to), the streaming reader will
treat the end of the stream as an unexpected EOF and error out.
Likewise, in async contexts where bytes can arrive in separate events,
trying to read chunk of Ion bytes can fail if the slice doesn't contain
a complete Ion stream.

This PR adds two new types: a `BinaryBuffer` and a `RawBinaryBufferReader`.

The `BinaryBuffer` is a stack-allocated wrapper around an implementation of
`AsRef<u8>`. It provides methods to read binary Ion encoding primitives like
`VarUInt`, `VarInt`, and `NOP` from the data source. It is very cheap to
create and does not modify the data source.

The `RawBinaryBufferReader` provides a full `Reader` interface over a
`BinaryBuffer`. If it encounters an unexpected EOF while reading, it returns
an `IonError::Incomplete`. Unlike the original `RawBinaryReader`, however,
it is guaranteed to still be in a good state afterwards. If more data
becomes available later on, it can be appended to the buffer and the
`RawBinaryBufferReader` can try parsing it again.

Finally, a new configuration for the `ion-tests` integration suite has been
added that exercises the `Reader` interface in the `RawBinaryBufferReader`,
demonstrating that it is spec-compliant.

This PR duplicates a fair amount of functionality that is available
elsewhere in the crate. Follow-on PRs will consolidate these implementations;
it is expected that the blocking raw binary reader will be re-implemented
as a wrapper around the non-blocking reader, making calls to `read` whenever
an `Incomplete` is encountered.