Optimizes reading binary symbol and integer values (#396)
This PR:

* Eliminates redundant boundary checking that happens when
integer and symbol values are read. It does this by offering
unchecked `UInt` decoding methods that read a single value from a
provided slice.
* Removes support for reading symbol IDs that require more than 8
bytes to encode. This limits the size of the symbol table to about
18 quintillion symbol values, which seems reasonable.
* Adds `Into<Integer>` implementations for Rust's integer primitives.

Together, these reduced the time needed to read every value in a
1.3GB Ion log file by about 6%.

Fixes #395.