The IonValueFormatter now uses the correct s-exp delimiter
Previously, the IonValueFormatter (which is used in the `Display`
impl for the `Element` API) would use commas to delimit child
values in an s-expression. Commas are the delimiter for list values
and struct fields, but s-expressions simply use whitespace.

Following this change, s-expression values are delimited with a
single space.

Fixes #403.