Removes dependency on `ion-c-sys` (#418)
This PR removes all code that was previously behind the `ion_c` feature
gate. Many of the native implementation APIs were developed with an eye
towards maintaining compatability with the `ion-c-sys` implementation.
However, this limited our ability to offer idiomatic APIs.

`ion-hash` continues to have a dependency on `ion-c-sys` for the time being,
but has been modified to depend on to the most recently published version of
`ion-rust` that supported it. A follow-on PR will update `ion-hash` to
depend on the native implementations instead.