Add functions to access Sign and Magnitude of Decimal
This change adds access to the Sign and unscaled Magnitude of a Decimal
in its inherent implementation.

I did not expose Coefficient as it was. That seemed to leak too much of
the implementation and not be very user friendly.

I chose to expose the Sign enum as it was: it seems clearest this way
and doesn't expose any implementation details which are likely to change.

I intentionally decoupled the unscaled magnitude from the stored
representation as I think it's more useful this way.