Removes uses of deprecated methods from `chrono` (#471)
Starting in v0.4.23, `chrono` deprecated a number of methods that would panic if the parameters were out of bounds (e.g. `month=400`) in favor of new versions that would return a `None` if the inputs were bad.

This PR replaces usages of the deprecated methods with their safe counterparts. Unfortunately, the safe counterparts are rather more verbose. However, most of the impacted code was in unit tests.